@@ -4,53 +4,97 @@ declare(strict_types=1);
 
 namespace spaceonfire\Type;
 
-use ArrayIterator;
-use IteratorAggregate;
-use Webmozart\Assert\Assert;
+use spaceonfire\Common\Factory\SingletonStorageTrait;
+use spaceonfire\Common\Factory\StaticConstructorInterface;
 
-abstract class AbstractAggregatedType implements TypeInterface, IteratorAggregate
+/**
+ * @implements \IteratorAggregate<TypeInterface>
+ */
+abstract class AbstractAggregatedType implements TypeInterface, StaticConstructorInterface, \IteratorAggregate
 {
+    use SingletonStorageTrait;
+
+    public const DELIMITER = '';
+
     /**
      * @var TypeInterface[]
      */
-    protected $types;
+    protected array $types;
 
     /**
-     * @var string
+     * @param TypeInterface[] $types
      */
-    protected $delimiter;
-
-    public function __construct(iterable $types, string $delimiter)
+    final private function __construct(array $types)
     {
-        $types = array_values($this->prepareTypes($types));
-
-        Assert::allIsInstanceOf($types, TypeInterface::class);
-        Assert::minCount($types, 2);
-
         $this->types = $types;
-        $this->delimiter = $delimiter;
+
+        self::singletonAttach($this);
     }
 
-    /**
-     * @inheritDoc
-     */
+    final public function __destruct()
+    {
+        self::singletonDetach($this);
+    }
+
     public function __toString(): string
     {
-        return implode($this->delimiter, $this->types);
+        return \implode(static::DELIMITER, $this->types);
     }
 
-    public function getIterator(): ArrayIterator
+    /**
+     * @param TypeInterface ...$types
+     * @return static
+     */
+    public static function new(TypeInterface ...$types): self
     {
-        return new ArrayIterator($this->types);
+        if (1 !== \strlen(static::DELIMITER)) {
+            throw new \LogicException(\sprintf('%s::DELIMITER should be 1 symbol string.', static::class));
+        }
+
+        $types = \array_values(self::prepareTypes($types));
+
+        if (2 > \count($types)) {
+            throw new \InvalidArgumentException(\sprintf(
+                '%s::new() requires at least 2 different types, %d given.',
+                static::class,
+                \count($types),
+            ));
+        }
+
+        return self::singletonFetch(self::singletonKey($types)) ?? new static($types);
     }
 
-    private function prepareTypes(iterable $types): array
+    /**
+     * @return \Traversable<TypeInterface>
+     */
+    public function getIterator(): \Traversable
+    {
+        return new \ArrayIterator($this->types);
+    }
+
+    /**
+     * @param iterable<TypeInterface> $value
+     * @return string
+     */
+    final protected static function singletonKey($value): string
+    {
+        return \implode(
+            static::DELIMITER,
+            $value instanceof \Traversable ? \iterator_to_array($value, false) : (array)$value
+        );
+    }
+
+    /**
+     * @param TypeInterface[] $types
+     * @return TypeInterface[]
+     */
+    private static function prepareTypes(iterable $types): array
     {
         $output = [];
 
         foreach ($types as $type) {
             if ($type instanceof static) {
-                $output += $this->prepareTypes($type);
+                $output += self::prepareTypes($type->types);
                 continue;
             }
 
