@@ -12,32 +12,19 @@ final class InstanceOfTypeFactory implements TypeFactoryInterface
 {
     use TypeFactoryTrait;
 
-    /**
-     * @var bool
-     */
-    private $autoload;
+    private bool $autoload;
 
-    /**
-     * InstanceOfTypeFactory constructor.
-     * @param bool $autoload
-     */
     public function __construct(bool $autoload = true)
     {
         $this->autoload = $autoload;
     }
 
-    /**
-     * @inheritDoc
-     */
     public function supports(string $type): bool
     {
         $type = $this->removeWhitespaces($type);
-        return class_exists($type, $this->autoload) || interface_exists($type, $this->autoload);
+        return \class_exists($type, $this->autoload) || \interface_exists($type, $this->autoload);
     }
 
-    /**
-     * @inheritDoc
-     */
     public function make(string $type): TypeInterface
     {
         $type = $this->removeWhitespaces($type);
@@ -46,6 +33,7 @@ final class InstanceOfTypeFactory implements TypeFactoryInterface
             throw new TypeNotSupportedException($type, InstanceOfType::class);
         }
 
-        return new InstanceOfType($type);
+        /** @phpstan-var class-string $type */
+        return InstanceOfType::new($type);
     }
 }
