@@ -6,15 +6,8 @@ namespace spaceonfire\Type\Factory;
 
 trait TypeFactoryTrait
 {
-    /**
-     * @var TypeFactoryInterface|null
-     */
-    protected $parent;
+    protected ?TypeFactoryInterface $parent = null;
 
-    /**
-     * Setter for `parent` property
-     * @param TypeFactoryInterface $parent
-     */
     public function setParent(TypeFactoryInterface $parent): void
     {
         $this->parent = $parent;
@@ -22,6 +15,6 @@ trait TypeFactoryTrait
 
     protected function removeWhitespaces(string $string): string
     {
-        return preg_replace('/\s+/', '', $string) ?? $string;
+        return \preg_replace('/\s+/', '', $string) ?? $string;
     }
 }
