@@ -12,25 +12,19 @@ final class MixedTypeFactory implements TypeFactoryInterface
 {
     use TypeFactoryTrait;
 
-    /**
-     * @inheritDoc
-     */
     public function supports(string $type): bool
     {
-        $type = $this->removeWhitespaces($type);
-        return MixedType::NAME === $type;
+        return MixedType::NAME === $this->removeWhitespaces($type);
     }
 
-    /**
-     * @inheritDoc
-     */
     public function make(string $type): TypeInterface
     {
         $type = $this->removeWhitespaces($type);
+
         if (!$this->supports($type)) {
             throw new TypeNotSupportedException($type, MixedType::class);
         }
 
-        return new MixedType();
+        return MixedType::new();
     }
 }
