@@ -8,12 +8,21 @@ use PHPUnit\Framework\TestCase;
 
 class DisjunctionTypeTest extends TestCase
 {
+    public function testFactoryOneArgument(): void
+    {
+        $this->expectException(\InvalidArgumentException::class);
+
+        DisjunctionType::new(
+            BuiltinType::new(BuiltinType::INT),
+        );
+    }
+
     public function testCheck(): void
     {
-        $type = new DisjunctionType([
-            new BuiltinType(BuiltinType::INT),
-            new BuiltinType(BuiltinType::NULL),
-        ]);
+        $type = DisjunctionType::new(
+            BuiltinType::new(BuiltinType::INT),
+            BuiltinType::new(BuiltinType::NULL),
+        );
         self::assertTrue($type->check(1));
         self::assertTrue($type->check(null));
         self::assertFalse($type->check('1'));
@@ -21,10 +30,10 @@ class DisjunctionTypeTest extends TestCase
 
     public function testStringify(): void
     {
-        $type = new DisjunctionType([
-            new BuiltinType(BuiltinType::INT),
-            new BuiltinType(BuiltinType::NULL),
-        ]);
+        $type = DisjunctionType::new(
+            BuiltinType::new(BuiltinType::INT),
+            BuiltinType::new(BuiltinType::NULL),
+        );
         self::assertSame('int|null', (string)$type);
     }
 }
