@@ -13,15 +13,13 @@ class CollectionTypeFactoryTest extends TestCase
     private function makeFactory(?TypeFactoryInterface $iterableTypeFactory = null): CollectionTypeFactory
     {
         $factory = new CollectionTypeFactory($iterableTypeFactory);
-        $factory->setParent(CompositeTypeFactory::makeWithDefaultFactories());
+        $factory->setParent(TypeFactoryAggregate::default());
         return $factory;
     }
 
     private function makeIterableTypeFactory(): TypeFactoryInterface
     {
-        return new PartialSupportTypeFactory(new BuiltinTypeFactory(), function (string $type): bool {
-            return in_array($type, ['array', 'iterable'], true);
-        });
+        return new PartialSupportTypeFactory(new BuiltinTypeFactory(), fn (string $type): bool => in_array($type, ['array', 'iterable'], true));
     }
 
     public function testSupports(): void
